requirements.txt
scanpy >= 1.10.0
numpy >= 1.23.1
anndata >= 0.10.6
flax >= "0.8.2"
ott-jax >= "0.4.6"
optax >= "0.2.1"
chex >= "0.1.81"
clu >= "0.0.11"
tqdm >= "4.66.1"